+++ client/src/features/editor/Editor.tsx
@@ -5,7 +5,7 @@ import { BlockLinkedList } from "@noctaCrdt/LinkedList";
 import { Block as CRDTBlock } from "@noctaCrdt/Node";
 import { BlockId } from "@noctaCrdt/NodeId";
 import { useRef, useState, useCallback, useEffect } from "react";
-import { Block } from "@src/features/editor/components/block/Block";
+import { Block } from "@src/features/editor/components/Block/Block";
 import { useMarkdownGrammer } from "@src/features/editor/hooks/useMarkdownGrammer";
 import { editorContainer, editorTitleContainer, editorTitle } from "./Editor.style";
 import { useBlockDragAndDrop } from "./hooks/useBlockDragAndDrop";
