+++ client/src/features/editor/components/IconBlock/IconBlock.tsx
